refactor: break schemas/sdk dependency cycle
schemas depended on sdk (runtime) only for validateSchema in its publish
scripts, while sdk depended on schemas for types — a cyclic workspace
dependency. Move validateSchema and its result types into
packages/schemas/validate (where schema validation belongs), and have the
sdk re-export them so its public API is unchanged.

Dependency graph is now one-directional: sdk -> schemas -> shared.
