+++ packages/schemas/scripts/publish-schema.ts
@@ -2,7 +2,7 @@
 import fs from 'node:fs'
 import path from 'node:path'
 import { fileURLToPath } from 'node:url'
-import { validateSchema } from '@ensmetadata/sdk'
+import { validateSchema } from '@ensmetadata/schemas/validate'
 import {
   type PublishedSchema,
   arg,
